feat: use typescript
This commit migrates the tokenlist codebase to typescript
