+++ .eslintignore
@@ -1,0 +1,4 @@
+/**/*.js
+build/**
+jest.config.ts
+
