+++ .prettierignore
@@ -1,0 +1,4 @@
+dist/
+build/
+**/*.json
+*.json
