+++ README.md
@@ -22,6 +22,7 @@ Token assets should be 256x256px PNG files.
 ### Wrapped token logo generation
 
 The wrapped token logo is generated by combining several images:
+
 1. the mask layer doubles up as a background and is drawn first
 2. the underlying token logo asset is drawn over the top, but masked by first layer (replaces colour channel values but reuses existing alpha channel)
 3. the overlay layer is drawn over the top, no masking applied
