+++ buttonwood.tokenlist.json
@@ -1,6 +1,6 @@
 {
   "name": "Buttonwood",
-  "timestamp": "2021-11-09T00:52:02.680Z",
+  "timestamp": "2021-11-09T21:21:45.081Z",
   "version": {
     "major": 0,
     "minor": 1,
